# api_service/app/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.DataResponse:
    properties:
      count:
        type: integer
      data: {}
    type: object
...
  models.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  models.IdResponse:
    properties:
      id:
        type: string
      message:
        type: string
    type: object
...
  models.PutgDTO:
    properties:
      coating:
        items:
          type: string
        type: array
      construction:
        items:
          $ref: '#/definitions/proto.PutgConstructions'
        type: array
      flangeId:
        type: string
      form:
        type: string
      graphite:
        items:
          type: string
        type: array
      iLimiter:
        $ref: '#/definitions/proto.PutgMaterials'
      mounting:
        items:
          type: string
        type: array
      oLimiter:
        $ref: '#/definitions/proto.PutgMaterials'
      obturator:
        $ref: '#/definitions/proto.PutgMaterials'
      reinforce:
        $ref: '#/definitions/proto.PutgMaterials'
      temperatures:
        items:
          $ref: '#/definitions/proto.PutgTemp'
        type: array
      typeFlId:
        type: string
      typePr:
        type: string
    required:
    - coating
    - construction
    - flangeId
    - form
    - graphite
    - mounting
    - temperatures
    - typeFlId
    - typePr
    type: object
  models.PutgImageDTO:
    properties:
      form:
        type: string
      gasket:
        type: string
      url:
        type: string
    required:
    - form
    - gasket
    - url
    type: object
...
  proto.ConTemp:
    properties:
      constructions:
        items:
          $ref: '#/definitions/proto.Constr'
        type: array
      temp:
        type: string
    type: object
  proto.Constr:
    properties:
      obturators:
        items:
          $ref: '#/definitions/proto.PutgObt'
        type: array
      short:
        type: string
    type: object
...
  proto.Putg:
    properties:
      coating:
        items:
          type: string
        type: array
      construction:
        items:
          $ref: '#/definitions/proto.PutgConstructions'
        type: array
      form:
        type: string
      graphite:
        items:
          type: string
        type: array
      iLimiter:
        $ref: '#/definitions/proto.PutgMaterials'
      id:
        type: string
      mounting:
        items:
          type: string
        type: array
      oLimiter:
        $ref: '#/definitions/proto.PutgMaterials'
      obturator:
        $ref: '#/definitions/proto.PutgMaterials'
      reinforce:
        $ref: '#/definitions/proto.PutgMaterials'
      temperatures:
        items:
          $ref: '#/definitions/proto.PutgTemp'
        type: array
      typeFlId:
        type: string
      typePr:
        type: string
    type: object
  proto.PutgConstructions:
    properties:
      grap:
        type: string
      temperatures:
        items:
          $ref: '#/definitions/proto.ConTemp'
        type: array
    type: object
  proto.PutgImage:
    properties:
      form:
        type: string
      gasket:
        type: string
      id:
        type: string
      url:
        type: string
    type: object
  proto.PutgMaterials:
    properties:
      default:
        type: string
      obturators:
        items:
          type: string
        type: array
      values:
        items:
          type: string
        type: array
    type: object
  proto.PutgObt:
    properties:
      imageUrl:
        type: string
      short:
        type: string
    type: object
  proto.PutgTemp:
    properties:
      grap:
        type: string
      temps:
        items:
          $ref: '#/definitions/proto.Temp'
        type: array
    type: object
...
  proto.Temp:
    properties:
      id:
        type: string
      mods:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Service for Sealur
  title: Sealur
  version: 0.14.0
paths:
...
  /sealur-pro/putg:
    get:
      consumes:
      - application/json
      description: получение прокладок путг
      parameters:
      - description: form
        in: query
        name: form
        required: true
        type: string
      - description: flange id
        in: query
        name: flangeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.DataResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/proto.Putg'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Putg
      tags:
      - Sealur Pro -> putg
    post:
      consumes:
      - application/json
      description: создание прокладки путг
      parameters:
      - description: putg info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.PutgDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Putg
      tags:
      - Sealur Pro -> putg
  /sealur-pro/putg-image:
    get:
      consumes:
      - application/json
      description: получение списка чертежей для путг
      parameters:
      - description: form
        in: query
        name: form
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.DataResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/proto.PutgImage'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Putg Image
      tags:
      - Sealur Pro -> putg-image
    post:
      consumes:
      - multipart/form-data
      - application/json
      description: создание чертежа путг
      parameters:
      - description: putg image info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.PutgImageDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Putg Image
      tags:
      - Sealur Pro -> putg-image
  /sealur-pro/putg-image/{id}:
    delete:
      consumes:
      - application/json
      description: удаление чертежа путг
      parameters:
      - description: putg image id
        in: path
        name: id
        required: true
        type: string
      - description: file name
        in: query
        name: file
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Putg Image
      tags:
      - Sealur Pro -> putg-image
    put:
      consumes:
      - application/json
      description: обновление чертежа путг
      parameters:
      - description: putg image id
        in: path
        name: id
        required: true
        type: string
      - description: putg image info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.PutgImageDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Putg Image
      tags:
      - Sealur Pro -> putg-image
  /sealur-pro/putg/{id}:
    delete:
      consumes:
      - application/json
      description: удаление прокладки путг
      parameters:
      - description: putg id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Putg
      tags:
      - Sealur Pro -> putg
    put:
      consumes:
      - application/json
      description: обновление прокладки путг
      parameters:
      - description: putg info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.PutgDTO'
      - description: putg id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Putg
      tags:
      - Sealur Pro -> putg
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"